fix: network error redirects to invalid screen (#773)
