+++ src/sagas/networkSettings.js
@@ -5,6 +5,7 @@ import hathorLib from '@hathor/wallet-lib';
 import { getGlobalWallet } from '../modules/wallet';
 import helpers from '../utils/helpers';
 import walletUtils from '../utils/wallet'
+import { t } from 'ttag';
 
 /**
  * Change network settings with new data
@@ -29,8 +30,6 @@ export function* changeNetworkSettings({ data, pin }) {
     hathorLib.config.getWalletServiceBaseWsUrl() :
     '';
 
-  const currentNetwork = wallet.getNetwork();
-
   // Update new server in storage and in the config singleton
   wallet.changeServer(data.node);
 
@@ -41,7 +40,17 @@ export function* changeNetworkSettings({ data, pin }) {
 
   let versionData;
   try {
-    versionData = yield call([wallet, wallet.getVersionData]);
+    const versionUrl = new URL('version', data.node).toString();
+    /*
+     * Using fetch instead of the wallet-lib axios instance to bypass
+     * the axios interceptor that is used to check for network errors.
+     * If this request fails, we don't want the network error modal to be shown.
+     */
+    const response = yield call(fetch, versionUrl);
+    if (!response.ok) {
+      throw new Error(`Network response was not ok: ${response.status}`);
+    }
+    versionData = yield call([response, response.json]);
   } catch (e) {
     // Invalid node, so go back to the previous server and return
     wallet.changeServer(currentServer);
