+++ src/screens/NetworkSettingsRecovery.js
@@ -1,0 +1,12 @@
+import React from 'react';
+import NetworkSettingsForm from '../components/NetworkSettingsForm';
+
+function NetworkSettingsRecovery(props) {
+  return (
+    <div className="content-wrapper">
+      <NetworkSettingsForm {...props} />
+    </div>
+  );
+}
+
+export default NetworkSettingsRecovery;
